requirements.txt
audtorch>=0.6.4
cxxfilt>=0.3.0
lief>=0.15.1
lightning==2.1.3
line_profiler>=4.1.2
llvmlite>=0.42.0
matplotlib>=3.8.2
nb_log>=12.6
networkx>=3.1
numpy==1.26.3
pandas>=2.2.2
psutil>=5.9.8
pygraphviz>=1.9
python_json_logger>=0.1.10
PyYAML>=6.0.1
PyYAML>=6.0.1
redis>=5.0.1
scikit_learn>=1.4.0
torch==2.1.2
tqdm>=4.66.1
openpyxl>=3.1.5
graphviz>=0.17
pybind11>=2.12
pydantic==1.10.14